v1.2 switched back to double

+++ src/main/kotlin/operator/Tester.kt
@@ -1,6 +1,6 @@
 package operator
 
-import util.readFloat
+import util.readDouble
 import util.readInt
 import kotlin.random.Random
 
@@ -32,10 +32,10 @@ class Tester(N: Int) : Operator {
                 val k = readInt()
                 if (coins[i] or coins[j] != 0) coins flip k
             }
-            "GenFlip" -> coins[i] = if (Random.nextDouble() < readFloat()) 1 else 0
+            "GenFlip" -> coins[i] = if (Random.nextDouble() < readDouble()) 1 else 0
             "Gen1Bit" -> {
-                val p = readFloat()
-                val q = readFloat()
+                val p = readDouble()
+                val q = readDouble()
                 if (coins[i] == 0) {
                     if (Random.nextDouble() < p) coins[i] = 1
                 } else
